requirements.txt
tweepy==4.14.0
openai==1.3.0
python-dotenv==1.0.0
schedule==1.2.0
flask==3.0.0
flask-login==0.6.3
flask-wtf==1.2.1
werkzeug==3.0.1 